+++ crates/mongodb-agent-common/src/query/mod.rs
@@ -1,4 +1,3 @@
-pub mod arguments;
 mod column_ref;
 mod constants;
 mod execute_query_request;