+++ crates/ndc-query-plan/src/lib.rs
@@ -1,18 +1,16 @@
+mod mutation_plan;
 mod plan_for_query_request;
 mod query_plan;
 mod type_system;
 pub mod vec_set;
 
+pub use mutation_plan::*;
 pub use plan_for_query_request::{
     plan_for_query_request,
     query_context::QueryContext,
     query_plan_error::QueryPlanError,
+    plan_for_mutation_request,
     type_annotated_field::{type_annotated_field, type_annotated_nested_field},
 };
-pub use query_plan::{
-    Aggregate, AggregateFunctionDefinition, ComparisonOperatorDefinition, ComparisonTarget,
-    ComparisonValue, ConnectorTypes, ExistsInCollection, Expression, Field, NestedArray,
-    NestedField, NestedObject, OrderBy, OrderByElement, OrderByTarget, Query, QueryPlan,
-    Relationship, Relationships, Scope, VariableSet, VariableTypes,
-};
+pub use query_plan::*;
 pub use type_system::{inline_object_types, ObjectType, Type};