+++ crates/ndc-query-plan/src/plan_for_query_request/mod.rs
@@ -1,4 +1,6 @@
 mod helpers;
+mod plan_for_arguments;
+mod plan_for_mutation_request;
 pub mod query_context;
 pub mod query_plan_error;
 mod query_plan_state;
@@ -21,10 +23,12 @@ use query_plan_state::QueryPlanInfo;
 
 use self::{
     helpers::{find_object_field, find_object_field_path, lookup_relationship},
+    plan_for_arguments::plan_for_arguments,
     query_context::QueryContext,
     query_plan_error::QueryPlanError,
     query_plan_state::QueryPlanState,
 };
+pub use self::plan_for_mutation_request::plan_for_mutation_request;
 
 type Result<T> = std::result::Result<T, QueryPlanError>;
 
@@ -33,6 +37,7 @@ pub fn plan_for_query_request<T: QueryContext>(
     request: QueryRequest,
 ) -> Result<QueryPlan<T>> {
     let mut plan_state = QueryPlanState::new(context, &request.collection_relationships);
+    let collection_info = context.find_collection(&request.collection)?;
     let collection_object_type = context.find_collection_object_type(&request.collection)?;
 
     let mut query = plan_for_query(
@@ -42,6 +47,12 @@ pub fn plan_for_query_request<T: QueryContext>(
         request.query,
     )?;
     query.scope = Some(Scope::Root);
+
+    let arguments = plan_for_arguments(
+        &mut plan_state,
+        &collection_info.arguments,
+        request.arguments,
+    )?;
 
     let QueryPlanInfo {
         unrelated_joins,
@@ -70,7 +81,7 @@ pub fn plan_for_query_request<T: QueryContext>(
 
     Ok(QueryPlan {
         collection: request.collection,
-        arguments: request.arguments,
+        arguments,
         query,
         variables,
         variable_types,
@@ -680,7 +691,7 @@ fn plan_for_exists<T: QueryContext>(
                 ..Default::default()
             };
 
-            let join_key = plan_state.register_unrelated_join(collection, arguments, join_query);
+            let join_key = plan_state.register_unrelated_join(collection, arguments, join_query)?;
 
             let in_collection = plan::ExistsInCollection::Unrelated {
                 unrelated_collection: join_key,