+++ crates/ndc-query-plan/src/query_plan.rs
@@ -3,9 +3,7 @@ use std::{collections::BTreeMap, fmt::Debug, iter};
 use derivative::Derivative;
 use indexmap::IndexMap;
 use itertools::Either;
-use ndc_models::{
-    Argument, OrderDirection, RelationshipArgument, RelationshipType, UnaryComparisonOperator,
-};
+use ndc_models::{self as ndc, OrderDirection, RelationshipType, UnaryComparisonOperator};
 
 use crate::{vec_set::VecSet, Type};
 
@@ -22,9 +20,9 @@ pub trait ConnectorTypes {
     PartialEq(bound = "T::ScalarType: PartialEq")
 )]
 pub struct QueryPlan<T: ConnectorTypes> {
-    pub collection: ndc_models::CollectionName,
+    pub collection: ndc::CollectionName,
     pub query: Query<T>,
-    pub arguments: BTreeMap<ndc_models::ArgumentName, Argument>,
+    pub arguments: BTreeMap<ndc::ArgumentName, Argument<T>>,
     pub variables: Option<Vec<VariableSet>>,
 
     /// Types for values from the `variables` map as inferred by usages in the query request. It is
@@ -44,9 +42,10 @@ impl<T: ConnectorTypes> QueryPlan<T> {
     }
 }
 
-pub type Relationships<T> = BTreeMap<ndc_models::RelationshipName, Relationship<T>>;
-pub type VariableSet = BTreeMap<ndc_models::VariableName, serde_json::Value>;
-pub type VariableTypes<T> = BTreeMap<ndc_models::VariableName, VecSet<Option<Type<T>>>>;
+pub type Arguments<T> = BTreeMap<ndc::ArgumentName, Argument<T>>;
+pub type Relationships<T> = BTreeMap<ndc::RelationshipName, Relationship<T>>;
+pub type VariableSet = BTreeMap<ndc::VariableName, serde_json::Value>;
+pub type VariableTypes<T> = BTreeMap<ndc::VariableName, VecSet<Type<T>>>;
 
 #[derive(Derivative)]
 #[derivative(
@@ -56,8 +55,8 @@ pub type VariableTypes<T> = BTreeMap<ndc_models::VariableName, VecSet<Option<Typ
     PartialEq(bound = "")
 )]
 pub struct Query<T: ConnectorTypes> {
-    pub aggregates: Option<IndexMap<ndc_models::FieldName, Aggregate<T>>>,
-    pub fields: Option<IndexMap<ndc_models::FieldName, Field<T>>>,
+    pub aggregates: Option<IndexMap<ndc::FieldName, Aggregate<T>>>,
+    pub fields: Option<IndexMap<ndc::FieldName, Field<T>>>,
     pub limit: Option<u32>,
     pub aggregates_limit: Option<u32>,
     pub offset: Option<u32>,
@@ -93,20 +92,67 @@ impl<T: ConnectorTypes> Query<T> {
 }
 
 #[derive(Derivative)]
+#[derivative(
+    Clone(bound = ""),
+    Debug(bound = ""),
+    PartialEq(bound = "T::ScalarType: PartialEq")
+)]
+pub enum Argument<T: ConnectorTypes> {
+    /// The argument is provided by reference to a variable
+    Variable {
+        name: ndc::VariableName,
+        argument_type: Type<T::ScalarType>,
+    },
+    /// The argument is provided as a literal value
+    Literal {
+        value: serde_json::Value,
+        argument_type: Type<T::ScalarType>,
+    },
+    /// The argument was a literal value that has been parsed as an [Expression]
+    Predicate { expression: Expression<T> },
+}
+
+#[derive(Derivative)]
 #[derivative(Clone(bound = ""), Debug(bound = ""), PartialEq(bound = ""))]
 pub struct Relationship<T: ConnectorTypes> {
-    pub column_mapping: BTreeMap<ndc_models::FieldName, ndc_models::FieldName>,
+    pub column_mapping: BTreeMap<ndc::FieldName, ndc::FieldName>,
     pub relationship_type: RelationshipType,
-    pub target_collection: ndc_models::CollectionName,
-    pub arguments: BTreeMap<ndc_models::ArgumentName, RelationshipArgument>,
+    pub target_collection: ndc::CollectionName,
+    pub arguments: BTreeMap<ndc::ArgumentName, RelationshipArgument<T>>,
     pub query: Query<T>,
+}
+
+#[derive(Derivative)]
+#[derivative(
+    Clone(bound = ""),
+    Debug(bound = ""),
+    PartialEq(bound = "T::ScalarType: PartialEq")
+)]
+pub enum RelationshipArgument<T: ConnectorTypes> {
+    /// The argument is provided by reference to a variable
+    Variable {
+        name: ndc::VariableName,
+        argument_type: Type<T::ScalarType>,
+    },
+    /// The argument is provided as a literal value
+    Literal {
+        value: serde_json::Value,
+        argument_type: Type<T::ScalarType>,
+    },
+    // The argument is provided based on a column of the source collection
+    Column {
+        name: ndc::FieldName,
+        argument_type: Type<T::ScalarType>,
+    },
+    /// The argument was a literal value that has been parsed as an [Expression]
+    Predicate { expression: Expression<T> },
 }
 
 #[derive(Derivative)]
 #[derivative(Clone(bound = ""), Debug(bound = ""), PartialEq(bound = ""))]
 pub struct UnrelatedJoin<T: ConnectorTypes> {
-    pub target_collection: ndc_models::CollectionName,
-    pub arguments: BTreeMap<ndc_models::ArgumentName, RelationshipArgument>,
+    pub target_collection: ndc::CollectionName,
+    pub arguments: BTreeMap<ndc::ArgumentName, RelationshipArgument<T>>,
     pub query: Query<T>,
 }
 
@@ -121,13 +167,13 @@ pub enum Scope {
 pub enum Aggregate<T: ConnectorTypes> {
     ColumnCount {
         /// The column to apply the count aggregate function to
-        column: ndc_models::FieldName,
+        column: ndc::FieldName,
         /// Whether or not only distinct items should be counted
         distinct: bool,
     },
     SingleColumn {
         /// The column to apply the aggregation function to
-        column: ndc_models::FieldName,
+        column: ndc::FieldName,
         /// Single column aggregate function name.
         function: T::AggregateFunction,
         result_type: Type<T::ScalarType>,
@@ -138,7 +184,7 @@ pub enum Aggregate<T: ConnectorTypes> {
 #[derive(Derivative)]
 #[derivative(Clone(bound = ""), Debug(bound = ""), PartialEq(bound = ""))]
 pub struct NestedObject<T: ConnectorTypes> {
-    pub fields: IndexMap<ndc_models::FieldName, Field<T>>,
+    pub fields: IndexMap<ndc::FieldName, Field<T>>,
 }
 
 #[derive(Derivative)]
@@ -158,7 +204,7 @@ pub enum NestedField<T: ConnectorTypes> {
 #[derivative(Clone(bound = ""), Debug(bound = ""), PartialEq(bound = ""))]
 pub enum Field<T: ConnectorTypes> {
     Column {
-        column: ndc_models::FieldName,
+        column: ndc::FieldName,
 
         /// When the type of the column is a (possibly-nullable) array or object,
         /// the caller can request a subset of the complete column data,
@@ -172,9 +218,9 @@ pub enum Field<T: ConnectorTypes> {
         /// The name of the relationship to follow for the subquery - this is the key in the
         /// [Query] relationships map in this module, it is **not** the key in the
         /// [ndc::QueryRequest] collection_relationships map.
-        relationship: ndc_models::RelationshipName,
-        aggregates: Option<IndexMap<ndc_models::FieldName, Aggregate<T>>>,
-        fields: Option<IndexMap<ndc_models::FieldName, Field<T>>>,
+        relationship: ndc::RelationshipName,
+        aggregates: Option<IndexMap<ndc::FieldName, Aggregate<T>>>,
+        fields: Option<IndexMap<ndc::FieldName, Field<T>>>,
     },
 }
 
@@ -274,34 +320,34 @@ pub struct OrderByElement<T: ConnectorTypes> {
 pub enum OrderByTarget<T: ConnectorTypes> {
     Column {
         /// The name of the column
-        name: ndc_models::FieldName,
+        name: ndc::FieldName,
 
         /// Path to a nested field within an object column
-        field_path: Option<Vec<ndc_models::FieldName>>,
+        field_path: Option<Vec<ndc::FieldName>>,
 
         /// Any relationships to traverse to reach this column. These are translated from
-        /// [ndc_models::OrderByElement] values in the [ndc_models::QueryRequest] to names of relation
+        /// [ndc::OrderByElement] values in the [ndc::QueryRequest] to names of relation
         /// fields for the [QueryPlan].
-        path: Vec<ndc_models::RelationshipName>,
+        path: Vec<ndc::RelationshipName>,
     },
     SingleColumnAggregate {
         /// The column to apply the aggregation function to
-        column: ndc_models::FieldName,
+        column: ndc::FieldName,
         /// Single column aggregate function name.
         function: T::AggregateFunction,
 
         result_type: Type<T::ScalarType>,
 
         /// Any relationships to traverse to reach this aggregate. These are translated from
-        /// [ndc_models::OrderByElement] values in the [ndc_models::QueryRequest] to names of relation
+        /// [ndc::OrderByElement] values in the [ndc::QueryRequest] to names of relation
         /// fields for the [QueryPlan].
-        path: Vec<ndc_models::RelationshipName>,
+        path: Vec<ndc::RelationshipName>,
     },
     StarCountAggregate {
         /// Any relationships to traverse to reach this aggregate. These are translated from
-        /// [ndc_models::OrderByElement] values in the [ndc_models::QueryRequest] to names of relation
+        /// [ndc::OrderByElement] values in the [ndc::QueryRequest] to names of relation
         /// fields for the [QueryPlan].
-        path: Vec<ndc_models::RelationshipName>,
+        path: Vec<ndc::RelationshipName>,
     },
 }
 
@@ -310,42 +356,42 @@ pub enum OrderByTarget<T: ConnectorTypes> {
 pub enum ComparisonTarget<T: ConnectorTypes> {
     Column {
         /// The name of the column
-        name: ndc_models::FieldName,
+        name: ndc::FieldName,
 
         /// Path to a nested field within an object column
-        field_path: Option<Vec<ndc_models::FieldName>>,
+        field_path: Option<Vec<ndc::FieldName>>,
 
         field_type: Type<T::ScalarType>,
 
         /// Any relationships to traverse to reach this column. These are translated from
-        /// [ndc_models::PathElement] values in the [ndc_models::QueryRequest] to names of relation
+        /// [ndc::PathElement] values in the [ndc::QueryRequest] to names of relation
         /// fields for the [QueryPlan].
-        path: Vec<ndc_models::RelationshipName>,
+        path: Vec<ndc::RelationshipName>,
     },
     ColumnInScope {
         /// The name of the column
-        name: ndc_models::FieldName,
+        name: ndc::FieldName,
 
         /// The named scope that identifies the collection to reference. This corresponds to the
         /// `scope` field of the [Query] type.
         scope: Scope,
 
         /// Path to a nested field within an object column
-        field_path: Option<Vec<ndc_models::FieldName>>,
+        field_path: Option<Vec<ndc::FieldName>>,
 
         field_type: Type<T::ScalarType>,
     },
 }
 
 impl<T: ConnectorTypes> ComparisonTarget<T> {
-    pub fn column_name(&self) -> &ndc_models::FieldName {
+    pub fn column_name(&self) -> &ndc::FieldName {
         match self {
             ComparisonTarget::Column { name, .. } => name,
             ComparisonTarget::ColumnInScope { name, .. } => name,
         }
     }
 
-    pub fn relationship_path(&self) -> &[ndc_models::RelationshipName] {
+    pub fn relationship_path(&self) -> &[ndc::RelationshipName] {
         match self {
             ComparisonTarget::Column { path, .. } => path,
             ComparisonTarget::ColumnInScope { .. } => &[],
@@ -373,7 +419,7 @@ pub enum ComparisonValue<T: ConnectorTypes> {
         value_type: Type<T::ScalarType>,
     },
     Variable {
-        name: ndc_models::VariableName,
+        name: ndc::VariableName,
         variable_type: Type<T::ScalarType>,
     },
 }
@@ -402,7 +448,7 @@ pub enum ExistsInCollection {
     Related {
         /// Key of the relation in the [Query] joins map. Relationships are scoped to the sub-query
         /// that defines the relation source.
-        relationship: ndc_models::RelationshipName,
+        relationship: ndc::RelationshipName,
     },
     Unrelated {
         /// Key of the relation in the [QueryPlan] joins map. Unrelated collections are not scoped