+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 This changelog documents the changes between release versions.
 
 ## [Unreleased]
+- Accept predicate arguments in native mutations and native queries ([#92](https://github.com/hasura/ndc-mongodb/pull/92))
 
 ## [1.0.0] - 2024-07-09